feature oriented architecture for Organisation domain
