+++ config/routes.yaml
@@ -5,10 +5,6 @@ index:
 fos_js_routing:
     resource: "@FOSJsRoutingBundle/Resources/config/routing/routing.xml"
 
-organisation:
-    resource: routes/organisations.yaml
-    prefix: organisations
-
 speaker:
     resource: routes/speakers.yaml
     prefix: speakers
