--- config/routes/organisations.yaml
@@ -1,42 +1,0 @@
-organisation_list:
-    path: /
-    controller: App\Controller\Organisation\Index::handle
-    methods: [GET]
-
-organisation_create:
-    path: /create
-    controller: App\Controller\Organisation\Create::handle
-    methods: [GET, POST]
-
-organisation_show:
-    path: /{id}
-    controller: App\Controller\Organisation\Show:handle
-    methods: [GET]
-    requirements:
-        id: '%routing.uuid%'
-
-organisation_edit:
-    path: /{id}/edit
-    controller: App\Controller\Organisation\Edit:handle
-    methods: [GET, PUT]
-    requirements:
-        id: '%routing.uuid%'
-
-organisation_delete:
-    path: /{id}/delete
-    controller: App\Controller\Organisation\Delete:handle
-    methods: [DELETE]
-    requirements:
-        id: '%routing.uuid%'
-
-organization_confirm_delete:
-    path: /{id}/confirm_delete
-    controller: App\Controller\Organisation\DeleteConfirmation:handle
-    methods: [GET]
-    requirements:
-        id: '%routing.uuid%'
-
-organisation_csv_upload:
-    path: /upload
-    controller: App\Controller\Organisation\Upload::handle
-    methods: [GET, POST]
