--- src/Config/Organisation.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Config;
-
-final class Organisation
-{
-    private $maxListOrganisationEntries;
-
-    public function __construct(int $maxListOrganisationEntries)
-    {
-        $this->maxListOrganisationEntries = $maxListOrganisationEntries;
-    }
-
-    public function getMaxListOrganisationEntries(): int
-    {
-        return $this->maxListOrganisationEntries;
-    }
-}
