--- src/Controller/Organisation/DeleteConfirmation.php
@@ -1,40 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Controller\Organisation;
-
-use App\Repository\Organisation\OrganisationRepositoryInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-use Twig\Environment as Twig;
-
-/**
- * @Security("is_granted('ROLE_ADMIN')")
- */
-final class DeleteConfirmation
-{
-    private $repository;
-    private $renderer;
-
-    public function __construct(
-        OrganisationRepositoryInterface $repository,
-        Twig $renderer
-    ) {
-        $this->repository = $repository;
-        $this->renderer = $renderer;
-    }
-
-    public function handle(string $id): Response
-    {
-        $organisation = $this->repository->find($id);
-        if (null === $organisation) {
-            throw new NotFoundHttpException();
-        }
-
-        return new Response($this->renderer->render('organisations/confirm_delete.html.twig', [
-           'organisation' => $organisation,
-        ]));
-    }
-}
