--- src/DataFixtures/OrganisationFixtures.php
@@ -1,58 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\DataFixtures;
-
-use App\Entity\Organisation;
-use App\Repository\Event\EventRepositoryInterface;
-use App\Repository\Organisation\OrganisationRepositoryInterface;
-use Doctrine\Bundle\FixturesBundle\Fixture;
-use Doctrine\Common\DataFixtures\DependentFixtureInterface;
-use Doctrine\Common\Persistence\ObjectManager;
-use Faker\Factory as Faker;
-
-final class OrganisationFixtures extends Fixture implements DependentFixtureInterface
-{
-    private $repository;
-    private $eventRepository;
-
-    public function __construct(OrganisationRepositoryInterface $repository, EventRepositoryInterface $eventRepository)
-    {
-        $this->repository = $repository;
-        $this->eventRepository = $eventRepository;
-    }
-
-    public function load(ObjectManager $manager): void
-    {
-        $events = $this->eventRepository->findAll();
-        $faker = Faker::create();
-
-        for ($i = 0; $i < 10; $i++) {
-            $uuid = $this->repository->nextIdentity();
-            $organisation = new Organisation(
-                $uuid,
-                $faker->company,
-                "http://{$faker->domainName}",
-                $this->getReference("contact-$i")
-            );
-
-            if ($faker->boolean) {
-                $organisation->setComment($faker->sentence);
-            }
-
-            $organisation->addSponsoredEvent($events[$faker->numberBetween(0, count($events) - 1)]);
-            $manager->persist($organisation);
-        }
-
-        $manager->flush();
-    }
-
-    public function getDependencies(): array
-    {
-        return [
-            EventFixture::class,
-            ContactFixtures::class,
-        ];
-    }
-}
