--- src/Form/OrganisationCsvUploadType.php
@@ -1,31 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Form;
-
-use App\Dto\OrganisationCsvUploadRequest;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\FileType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class OrganisationCsvUploadType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('name', FileType::class, [
-                'label' => 'CSV file to upload',
-                ]
-            )
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => OrganisationCsvUploadRequest::class,
-        ]);
-    }
-}
