+++ src/Organisation/OrganisationFormType.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Organisation;
+
+use App\Form\ContactType;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\UrlType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+final class OrganisationFormType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('name', TextType::class, ['required' => true])
+            ->add('website', UrlType::class, [
+                'required' => true,
+                'help' => 'Must start with http:// or https://',
+            ])
+            ->add('contact', ContactType::class, ['required' => true])
+            ->add('comment', TextareaType::class, ['required' => false])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => OrganisationRequest::class,
+        ]);
+    }
+}
