+++ src/Organisation/Resources/config/doctrine/Organisation.orm.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="App\Organisation\Organisation">
+        <id name="id" type="guid">
+            <generator strategy="NONE"/>
+        </id>
+        <field name="name" length="255"/>
+        <field name="website" length="255"/>
+        <many-to-one field="contact" target-entity="App\Entity\Contact">
+            <cascade>
+                <cascade-persist/>
+            </cascade>
+        </many-to-one>
+        <field name="comment" type="text" nullable="true"/>
+        <many-to-many field="events" target-entity="App\Entity\Event">
+            <cascade>
+                <cascade-remove/>
+            </cascade>
+        </many-to-many>
+    </entity>
+</doctrine-mapping>
