+++ src/Organisation/Resources/config/packages/doctrine.yaml
@@ -1,0 +1,9 @@
+doctrine:
+  orm:
+    mappings:
+      App\Organisation:
+        is_bundle: false
+        type: xml
+        dir: '%kernel.project_dir%/src/Organisation/Resources/config/doctrine'
+        prefix: 'App\Organisation'
+        alias: App\Organisation
