+++ src/Organisation/Resources/config/routes.yaml
@@ -1,0 +1,3 @@
+organisation:
+  resource: routes/organisations.yaml
+  prefix: organisations
